minggu 6

+++ application/controllers/article.php
@@ -5,8 +5,8 @@ class article extends CI_Controller {
     function __construct(){
         parent::__construct();
 
-        $this->load->model("ModelArticle", "article");
-    
+        $this->load->model("modelarticle", 'article');
+        $this->load->model("ModelUser",'user');
     }
     
     
@@ -15,6 +15,10 @@ class article extends CI_Controller {
 	}
 
     public function new() {
+        if ($this->session->userdata('logged_in') ==NULL) {
+            redirect('/');
+        }
+
         $this->load->helper('form');
         $this->load->library('form_validation');
 
@@ -49,6 +53,10 @@ class article extends CI_Controller {
     }
 
     public function edit($articleID) {
+        if ($this->session->userdata('logged_in') ==NULL) {
+            redirect('/');
+        }
+
         $this->load->helper('form');
         $this->load->library('form_validation');
 
@@ -79,8 +87,50 @@ class article extends CI_Controller {
     }
 
     public function delete($articleID) {
+        if ($this->session->userdata('logged_in') ==NULL) {
+            redirect('/');
+        }
+        
         $this->article->delete($articleID);
 
         redirect();
     }
+
+    public function login() {
+        $this->load->helper('form');
+        $this->load->library('form_validation');
+
+        $this->form_validation->set_rules('username','Username', 'required');
+        $this->form_validation->set_rules('password','Password', 'required');
+    
+
+        if ($this->form_validation->run() == FALSE) {
+            return $this->load->view('login_page');
+        }
+
+        
+        try {
+            $user = $this->user->getByUsername($this->input->post('username'), false);
+            if (!password_verify($this->input->post('password'), $user[0]->password)) {
+                echo "password doesn't match";
+
+                return;
+            }
+
+            $this->session->set_userdata(array('userid' => $user[0]->id, 'logged_in' => true));
+            
+            redirect('/');
+        } catch (Exception $e) {
+            show_error($e->getMessage());
+            return;
+        }
+    }
+
+    public function logout() {
+        $this->session->sess_destroy();
+
+        redirect("/");
+    }
+    
+
 }
